+++ docs/changelog.md
@@ -1,9 +1,17 @@
 ## Pre-release Versions
 
+### v0.2.2
+
+- **[Update]** Internal improvements and bug fixes.
+- **[Examples]** Added two more examples:
+  - `(dice:)` A port of the CMFSC2 macro.
+  - `(article:)`/`(set-article:)` A port of the CMFSC2 macro.
+- **[Meta]** Demo and docs website.
+
 ### v0.2.1
 
 - **[Update]** Fixed bug in `Harlowe.variable()`.
-- **[New]** Added two examples:
+- **[Examples]** Added two examples:
   - `(textbox:)` A basic text input implementation.
   - `(hotkey:)` A basic hotkey / click link with keypress implementation.
 - **[Docs]** Docs typos.